backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
redis>=4.5.3,<6.0.0
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# File Handling
Pillow==10.1.0
python-magic==0.4.27

# Webhook & HTTP Requests
requests==2.31.0
celery==5.3.4

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0

# Environment Management
python-decouple==3.8
django-environ==0.11.2

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring & Logging
sentry-sdk==1.38.0
structlog==23.2.0

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# Validation
marshmallow==3.20.1
phonenumbers==8.13.24

# Background Tasks (using celery instead for compatibility)
# django-q==1.3.9  # Conflicts with newer redis versions

# Health Checks
django-health-check==3.17.0

# Model Utilities
django-model-utils==4.3.1

# UUID Support
shortuuid==1.0.11 